Only verify those commits made by the user.name.
We only want to verify commits that the user has made, not any commits
by any other collaboraters in the repository. This filters the commits
to only include those made by the user with name from git config
user.name.

In addition this also changes the config to look for ghostwriter.email
instead of author, this is a breaking change but is way less ambiguous.